Spring data modificar Map<> y Solucionar conexion a MQTT

+++ src/main/java/sri/project/sri_project/service/EstadisticasService.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public interface EstadisticasService {
 
-    Map<String, Double> obtenerDatosHumedadHistorica();
+    Map<String, Integer> obtenerDatosHumedadHistorica();
     Map<String, Integer> obtenerDatosModosRiego();
 
 }
